Move non-app router code to lib folder
