+++ apps/dashboard/src/app/(routes)/api/auth/logout/route.ts
@@ -1,3 +1,3 @@
-import { auth } from "@/auth"
+import { auth } from "@/lib/auth"
 
 export const GET = auth.logout
