+++ apps/dashboard/src/app/QueryProvider.tsx
@@ -1,13 +1,13 @@
 "use client"
 
-import { env } from "@/env"
+import { env } from "@/lib/env"
+import { TRPCProvider } from "@/lib/trpc"
 import type { AppRouter } from "@dotkomonline/gateway-trpc"
 import { useSession } from "@dotkomonline/oauth2/react"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { type CreateTRPCClientOptions, createTRPCClient, httpBatchLink, loggerLink } from "@trpc/client"
 import { type PropsWithChildren, useState } from "react"
 import superjson from "superjson"
-import { TRPCProvider } from "../trpc"
 
 export const QueryProvider = ({ children }: PropsWithChildren) => {
   const session = useSession()
