+++ apps/dashboard/src/app/article/queries.ts
@@ -1,9 +1,12 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 import { useQuery } from "@tanstack/react-query"
 
 export const useArticleAllQuery = () => {
   const trpc = useTRPC()
-  const { data: articles, ...query } = useQuery({ ...trpc.article.all.queryOptions({ take: 999 }), initialData: [] })
+  const { data: articles, ...query } = useQuery({
+    ...trpc.article.all.queryOptions({ take: 999 }),
+    initialData: [],
+  })
 
   return { articles, ...query }
 }
