+++ apps/dashboard/src/app/company/mutations.ts
@@ -1,7 +1,7 @@
 import { useRouter } from "next/navigation"
 
-import { useQueryNotification } from "@/notifications"
-import { useTRPC } from "@/trpc"
+import { useQueryNotification } from "@/lib/notifications"
+import { useTRPC } from "@/lib/trpc"
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 
 export const useCreateCompanyMutation = () => {
