+++ apps/dashboard/src/app/event/mutations.ts
@@ -1,5 +1,5 @@
-import { useQueryGenericMutationNotification, useQueryNotification } from "@/notifications"
-import { useTRPC } from "@/trpc"
+import { useQueryGenericMutationNotification, useQueryNotification } from "@/lib/notifications"
+import { useTRPC } from "@/lib/trpc"
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 import { useRouter } from "next/navigation"
