+++ apps/dashboard/src/app/interest-group/mutations/use-delete-interest-group-mutation.ts
@@ -1,5 +1,5 @@
-import { useQueryNotification } from "@/notifications"
-import { useTRPC } from "@/trpc"
+import { useQueryNotification } from "@/lib/notifications"
+import { useTRPC } from "@/lib/trpc"
 
 import { useMutation, useQueryClient } from "@tanstack/react-query"
