+++ apps/dashboard/src/app/interest-group/queries/use-interest-group-all-query.ts
@@ -1,4 +1,4 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 
 import { useQuery } from "@tanstack/react-query"
