+++ apps/dashboard/src/app/job-listing/queries/use-job-listing-all-query.ts
@@ -1,4 +1,4 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 import { useQuery } from "@tanstack/react-query"
 
 export const useJobListingAllQuery = () => {
