+++ apps/dashboard/src/app/layout.tsx
@@ -4,11 +4,11 @@ import "@mantine/dates/styles.css"
 import "@mantine/notifications/styles.css"
 import "@mdxeditor/editor/style.css"
 import { Login } from "@/components/views/Login"
+import { auth } from "@/lib/auth"
 import { SessionProvider } from "@dotkomonline/oauth2/react"
 import type { Session } from "@dotkomonline/oauth2/session"
 import { Notifications } from "@mantine/notifications"
 import type { PropsWithChildren } from "react"
-import { auth } from "../auth"
 import { ApplicationShell } from "./ApplicationShell"
 import { ModalProvider } from "./ModalProvider"
 import { QueryProvider } from "./QueryProvider"
