+++ apps/dashboard/src/app/offline/queries/use-offlines-all-query.ts
@@ -1,4 +1,4 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 import { useQuery } from "@tanstack/react-query"
 
 export const useOfflineAllQuery = () => {
