+++ apps/dashboard/src/app/punishment/queries/use-count-users-with-mark-query.ts
@@ -1,4 +1,4 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 import type { MarkId } from "@dotkomonline/types"
 
 import { useQuery } from "@tanstack/react-query"
