+++ apps/dashboard/src/app/punishment/queries/use-personal-mark-get-by-mark-id.ts
@@ -1,4 +1,4 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 import type { MarkId } from "@dotkomonline/types"
 import { useQuery } from "@tanstack/react-query"
