+++ apps/dashboard/src/app/punishment/queries/use-punishment-all-query.ts
@@ -1,5 +1,4 @@
-import { useTRPC } from "@/trpc"
-
+import { useTRPC } from "@/lib/trpc"
 import { useQuery } from "@tanstack/react-query"
 
 export const usePunishmentAllQuery = () => {
