+++ apps/dashboard/src/app/user/components/UserSearch/queries.ts
@@ -1,4 +1,4 @@
-import { useTRPC } from "@/trpc"
+import { useTRPC } from "@/lib/trpc"
 import type { AttendanceId } from "@dotkomonline/types"
 
 import { useQuery } from "@tanstack/react-query"
