+++ apps/dashboard/src/lib/auth.ts
