+++ apps/dashboard/src/modules/article/queries/use-tags-all-query.ts
@@ -1,5 +1,4 @@
-import { useTRPC } from "../../../trpc"
-
+import { useTRPC } from "@/lib/trpc"
 import { useQuery } from "@tanstack/react-query"
 
 export const useTagsAllQuery = () => {
